Patch for issue #40 ORDER BY does not work.
There is the IndexReaderWide#prepareNext method that is responsible for preparing filter and reading data from Cassandra memtable and disk.
This method is being called on each call to IndexReader.hasNext() and sometimes to IndexReader.next() methods. IndexReader class implements
the UnfilteredPartitionIterator interface to allow Cassandra to retrieve data from unfiltered partitions.
The root cause is hardcoded false value when preparing ClusteringIndexNamesFilter for querying data from disks and memtables.
This patch changes this behaviour to use ReadCommand.isReversed() method which returns true when ORDER BY DESC clause is presented in the query.